requirements.txt
asgiref==3.9.1
Django==4.2.24
exceptiongroup==1.3.0
iniconfig==2.1.0
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
